# docs/02-design/openapi-campusswap.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CampusSwap 文档管理平台 API",
    "version": "1.0.0-m3",
    "description": "系统与权限域（M3 已实现，25 条）。字段口径与 docs/02-design/API_SPECIFICATION.md 逐字一致；所有 ID 在 JSON 中为字符串，时间为 yyyy-MM-dd HH:mm:ss。文档域 30 条随 M4 实现追加。"
  },
  "servers": [
    { "url": "http://localhost:10087", "description": "本机 dev（Spring Boot 4.1.1，端口 10087）" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "登录",
        "description": "公开接口。用户名不存在与密码错误返回完全一致的 401 文案（防用户名枚举）。成功时签发 token（UUID 32 位，Redis 存 2 小时）并预写权限缓存。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDtoReq" },
              "example": { "username": "admin", "password": "Admin@123" }
            }
          }
        },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultLoginVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/UserDisabled" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["认证"],
        "summary": "登出（幂等）",
        "description": "携带 token 即 DEL 掉该 token；重复登出仍返回 200，未带请求头返回 401。",
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["认证"],
        "summary": "获取当前登录用户",
        "description": "登录即可（不校验权限点）。权限码优先读 Redis 缓存，未命中回源数据库重建。",
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultUserInfoVo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/UserDisabled" }
        }
      }
    },
...
    "/api/permissions/tree": {
      "get": {
        "tags": ["权限管理"],
        "summary": "权限树（全量，不分页）",
        "description": "权限点 sys:perm。三层 DIR→MENU→BUTTON，一次查全后在内存组树（SQL 预算 1 条）。",
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultListPermissionVo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" }
        }
      }
    },
    "/api/permissions": {
      "post": {
        "tags": ["权限管理"],
        "summary": "新增权限节点",
        "description": "权限点 sys:perm:add。层级：目录的上级只能是根，菜单的上级必须是目录，按钮的上级必须是菜单；ancestors 由后端计算。",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PermissionCreateDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultPermissionVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/permissions/{id}": {
      "put": {
        "tags": ["权限管理"],
        "summary": "编辑权限节点",
        "description": "权限点 sys:perm:edit。支持子树整体移动并级联重写子孙 ancestors；禁止移到自身或自身子孙下（400）。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PermissionUpdateDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultPermissionVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["权限管理"],
        "summary": "删除权限节点",
        "description": "权限点 sys:perm:delete。有子节点或被角色/用户直授引用时 409。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/depts/tree": {
      "get": {
        "tags": ["部门管理"],
        "summary": "部门树（全量，不分页）",
        "description": "权限点 sys:dept。一次查全后在内存组树（SQL 预算 1 条）；同级按 sortOrder 升序、相同则按 id 升序。",
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultListDeptVo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" }
        }
      }
    },
    "/api/depts": {
      "post": {
        "tags": ["部门管理"],
        "summary": "新增部门",
        "description": "权限点 sys:dept:add。同一父节点下不可重名；ancestors 由后端计算。",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeptCreateDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultDeptVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" }
        }
      }
    },
    "/api/depts/{id}": {
      "put": {
        "tags": ["部门管理"],
        "summary": "编辑部门",
        "description": "权限点 sys:dept:edit。支持移动并级联重写子孙 ancestors；禁止移到自身或自身子孙下。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeptUpdateDtoReq" } } } },
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultDeptVo" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["部门管理"],
        "summary": "删除部门",
        "description": "权限点 sys:dept:delete。存在子部门或该部门下仍有未删除用户时 409。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/depts/{id}/roles": {
      "get": {
        "tags": ["部门管理"],
        "summary": "查询部门角色",
        "description": "权限点 sys:dept。返回部门已绑定的角色 ID 集合（SQL 预算 2 条）。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultDeptRoleVo" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["部门管理"],
        "summary": "部门绑定角色（覆盖式）",
        "description": "权限点 sys:role:grant。roleIds 是角色【ID】数组（与用户接口的 roles 编码数组不同）；保存后清该部门用户权限缓存。",
        "parameters": [{ "$ref": "#/components/parameters/PathId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeptRoleDtoReq" }, "example": { "roleIds": ["1", "2"] } } } },
        "responses": {
          "200": { "$ref": "#/components/responses/OkVoid" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/NoPermission" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "登录返回的 token，格式：Authorization: Bearer <token>" }
    },
    "parameters": {
      "PathId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^\\d+$" }, "description": "字符串形式的数字 ID" }
    },
    "responses": {
      "OkVoid": { "description": "成功（data 为 null）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" } } } },
      "BadRequest": { "description": "参数校验失败 / 业务规则拒绝（message 为中文原文）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 400, "message": "页码不能超过100，请缩小筛选范围后再试", "data": null } } } },
      "Unauthorized": { "description": "未登录或 token 失效", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 401, "message": "登录状态已失效，请重新登录", "data": null } } } },
      "NoPermission": { "description": "已登录但缺权限点", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 403, "message": "无权限执行该操作", "data": null } } } },
      "UserDisabled": { "description": "账号非 ACTIVE", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 403, "message": "账号已停用或冻结，请联系系统管理员", "data": null } } } },
      "NotFound": { "description": "资源不存在", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 404, "message": "用户不存在或已被删除", "data": null } } } },
      "Conflict": { "description": "唯一约束或状态冲突", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultVoid" }, "example": { "code": 409, "message": "登录名已存在，请更换", "data": null } } } }
    },
    "schemas": {
      "ResponseResultVoid": { "type": "object", "properties": { "code": { "type": "integer", "example": 200 }, "message": { "type": "string", "example": "成功" }, "data": { "nullable": true } } },
      "ResponseResultLoginVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/LoginVo" } } },
      "ResponseResultUserInfoVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/UserInfoVo" } } },
...
      "ResponseResultPermissionVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/PermissionVo" } } },
      "ResponseResultDeptVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/DeptVo" } } },
...
      "ResponseResultDeptRoleVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/DeptRoleVo" } } },
      "ResponseResultListPermissionVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/PermissionVo" } } } },
      "ResponseResultListDeptVo": { "type": "object", "properties": { "code": { "type": "integer" }, "message": { "type": "string" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/DeptVo" } } } },
...
      "AuditVo": {
        "type": "object",
        "description": "实体型 VO 的公共审计字段（时间与 ID 均为字符串）",
        "properties": {
          "createdAt": { "type": "string", "example": "2026-09-21 14:20:05" },
          "createdBy": { "type": "string", "example": "1" },
          "updatedAt": { "type": "string", "example": "2026-09-21 14:20:05" },
          "updatedBy": { "type": "string", "example": "1" }
        }
      },
      "LoginVo": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "32 位随机串，有效期 2 小时" },
          "userInfo": { "$ref": "#/components/schemas/UserInfoVo" },
          "roles": { "type": "array", "items": { "type": "string" } },
          "permissions": { "type": "array", "items": { "type": "string" } }
        }
      },
      "UserInfoVo": {
        "allOf": [{ "$ref": "#/components/schemas/AuditVo" }],
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "username": { "type": "string" }, "realName": { "type": "string" },
          "deptId": { "type": "string" }, "deptName": { "type": "string" },
          "roles": { "type": "array", "items": { "type": "string" } },
          "avatarUrl": { "type": "string", "nullable": true },
          "permissions": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "DeptRoleVo": {
        "type": "object",
        "properties": { "deptId": { "type": "string" }, "roleIds": { "type": "array", "items": { "type": "string" } } }
      },
      "PermissionVo": {
        "allOf": [{ "$ref": "#/components/schemas/AuditVo" }],
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "name": { "type": "string" }, "code": { "type": "string" },
          "type": { "type": "string", "enum": ["DIR", "MENU", "BUTTON"] },
          "parentId": { "type": "string" }, "ancestors": { "type": "string" },
          "path": { "type": "string", "nullable": true }, "icon": { "type": "string", "nullable": true },
          "sortOrder": { "type": "integer" },
          "children": { "type": "array", "items": { "$ref": "#/components/schemas/PermissionVo" } }
        }
      },
      "DeptVo": {
        "allOf": [{ "$ref": "#/components/schemas/AuditVo" }],
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "name": { "type": "string" },
          "parentId": { "type": "string" }, "ancestors": { "type": "string" },
          "sortOrder": { "type": "integer" },
          "children": { "type": "array", "items": { "$ref": "#/components/schemas/DeptVo" } }
        }
      },
      "LoginDtoReq": {
        "type": "object", "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 4, "maxLength": 64, "description": "用户名不能为空，且长度需在4到64之间" },
          "password": { "type": "string", "minLength": 6, "maxLength": 64, "description": "密码不能为空，且长度需在6到64之间（不 trim）" }
        }
      },
...
      "PermissionCreateDtoReq": {
        "type": "object", "required": ["name", "code", "type", "parentId"],
        "properties": {
          "name": { "type": "string", "maxLength": 64 },
          "code": { "type": "string", "pattern": "^[a-z0-9:]{2,64}$", "description": "权限编码只能包含小写字母、数字与冒号" },
          "type": { "type": "string", "enum": ["DIR", "MENU", "BUTTON"] },
          "parentId": { "type": "string", "description": "父节点 ID，根节点传 \"0\"" },
          "path": { "type": "string", "maxLength": 128, "description": "目录与菜单必填" },
          "icon": { "type": "string", "maxLength": 64 },
          "sortOrder": { "type": "integer", "minimum": 0, "default": 0 }
        }
      },
      "PermissionUpdateDtoReq": {
        "type": "object", "required": ["name", "type", "parentId"],
        "properties": {
          "name": { "type": "string", "maxLength": 64 },
          "type": { "type": "string", "enum": ["DIR", "MENU", "BUTTON"] },
          "parentId": { "type": "string", "description": "不能是自己或自己的子孙" },
          "path": { "type": "string", "maxLength": 128 },
          "icon": { "type": "string", "maxLength": 64 },
          "sortOrder": { "type": "integer", "minimum": 0 }
        }
      },
      "DeptCreateDtoReq": {
        "type": "object", "required": ["name", "parentId"],
        "properties": {
          "name": { "type": "string", "maxLength": 64, "description": "同一父节点下不可重名" },
          "parentId": { "type": "string", "description": "上级部门 ID，根部门传 \"0\"" },
          "sortOrder": { "type": "integer", "minimum": 0, "default": 0 }
        }
      },
      "DeptUpdateDtoReq": {
        "type": "object", "required": ["name", "parentId"],
        "properties": {
          "name": { "type": "string", "maxLength": 64 },
          "parentId": { "type": "string", "description": "不能是自己或自己的子孙" },
          "sortOrder": { "type": "integer", "minimum": 0 }
        }
      },
      "DeptRoleDtoReq": {
        "type": "object", "required": ["roleIds"],
        "properties": { "roleIds": { "type": "array", "items": { "type": "string" }, "description": "角色【ID】数组；空数组 = 清空绑定（注意与用户接口的 roles 编码数组区分）" } }
      }
    }
  }
}